+++ packages/cloud_firestore/cloud_firestore/lib/src/load_bundle_task.dart
@@ -1,0 +1,12 @@
+part of cloud_firestore;
+
+class LoadBundleTask {
+  LoadBundleTask._(this._delegate) {
+    LoadBundleTaskPlatform.verifyExtends(_delegate);
+  }
+
+  final LoadBundleTaskPlatform _delegate;
+
+  late final Stream<LoadBundleTaskSnapshot> stream =
+      _delegate.stream.map((event) => LoadBundleTaskSnapshot._(event));
+}