+++ packages/cloud_firestore/cloud_firestore_web/lib/src/interop/firestore.dart
@@ -6,6 +6,7 @@
 // ignore_for_file: prefer_void_to_null
 
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:firebase_core_web/firebase_core_web_interop.dart'
     hide jsify, dartify;
@@ -103,6 +104,79 @@ class Firestore extends JsObjectWrapper<firestore_interop.FirestoreJsImpl> {
 
   Future<Null> waitForPendingWrites() =>
       handleThenable(jsObject.waitForPendingWrites());
+
+  LoadBundleTask loadBundle(Uint8List bundle) {
+    return LoadBundleTask.getInstance(jsObject.loadBundle(bundle));
+  }
+
+  Future<Query> namedQuery(String name) async {
+    firestore_interop.QueryJsImpl query =
+        await handleThenable(jsObject.namedQuery(name));
+    return Query.fromJsObject(query);
+  }
+}
+
+class LoadBundleTask
+    extends JsObjectWrapper<firestore_interop.LoadBundleTaskJsImpl> {
+  LoadBundleTask._fromJsObject(firestore_interop.LoadBundleTaskJsImpl jsObject)
+      : super.fromJsObject(jsObject);
+
+  static final _expando = Expando<LoadBundleTask>();
+
+  /// Creates a new LoadBundleTask from a [jsObject].
+  static LoadBundleTask getInstance(
+    firestore_interop.LoadBundleTaskJsImpl jsObject,
+  ) {
+    return _expando[jsObject] ??= LoadBundleTask._fromJsObject(jsObject);
+  }
+
+  ///Tracks progress of loadBundle snapshots as the documents are loaded into cache
+  Stream<LoadBundleTaskProgress> stream() {
+    late StreamController<LoadBundleTaskProgress> controller;
+    controller =
+        StreamController<LoadBundleTaskProgress>.broadcast(onListen: () {
+      /// Calls underlying onProgress method on a LoadBundleTask [jsObject].
+      jsObject.onProgress(
+          allowInterop((firestore_interop.LoadBundleTaskProgressJsImpl data) {
+        controller.add(LoadBundleTaskProgress._fromJsObject(data));
+      }));
+      //todo(russellwheatley): implement error handling for loadBundle
+      // jsObject.JS$catch(allowInterop((firestore_interop.FirestoreError error) {
+      //   controller.addError(error);
+      // }));
+    }, onCancel: () {
+      controller.close();
+    });
+
+    return controller.stream;
+  }
+}
+
+class LoadBundleTaskProgress
+    extends JsObjectWrapper<firestore_interop.LoadBundleTaskProgressJsImpl> {
+  LoadBundleTaskProgress._fromJsObject(
+    firestore_interop.LoadBundleTaskProgressJsImpl jsObject,
+  ) : super.fromJsObject(jsObject);
+
+  static final _expando = Expando<LoadBundleTaskProgress>();
+
+  /// Creates a new LoadBundleTaskProgress from a [jsObject].
+  static LoadBundleTaskProgress getInstance(
+    firestore_interop.LoadBundleTaskProgressJsImpl jsObject,
+  ) {
+    return _expando[jsObject] ??=
+        LoadBundleTaskProgress._fromJsObject(jsObject);
+  }
+
+  int get bytesLoaded => int.parse(jsObject.bytesLoaded);
+
+  int get documentsLoaded => jsObject.documentsLoaded;
+
+  String get taskState => jsObject.taskState;
+
+  int get totalBytes => int.parse(jsObject.totalBytes);
+
+  int get totalDocuments => jsObject.totalDocuments;
 }
 
 class WriteBatch extends JsObjectWrapper<firestore_interop.WriteBatchJsImpl>