+++ packages/firebase_app_check/firebase_app_check_web/lib/src/interop/firebase_interop.dart
@@ -1,0 +1,14 @@
+// Copyright 2021, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// ignore_for_file: public_member_api_docs
+
+@JS('firebase')
+library firebase_interop.firebase;
+
+import 'package:js/js.dart';
+import 'app_check_interop.dart';
+
+@JS()
+external AppCheckJsImpl appCheck();