+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.0.5
+
+ - Update a dependency to the latest release.
+
 ## 5.0.4
 
  - **DOCS**: Add missing homepage/repository links (#6054).