+++ packages/firebase_dynamic_links/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.4
+
+ - Update a dependency to the latest release.
+
 ## 2.0.3
 
  - **DOCS**: Add missing homepage/repository links (#6054).